# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/keranjang": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan semua item di keranjang user dengan status active",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keranjang"
                ],
                "summary": "Dapatkan item keranjang",
                "responses": {
                    "200": {
                        "description": "Daftar item keranjang",
                        "schema": {
                            "$ref": "#/definitions/controllers.CartSummaryResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data keranjang",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menambahkan produk ke keranjang belanja user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keranjang"
                ],
                "summary": "Tambah produk ke keranjang",
                "parameters": [
                    {
                        "description": "Data produk yang akan ditambahkan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateKeranjangRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Produk berhasil ditambahkan ke keranjang",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Produk tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Stok produk tidak mencukupi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Gagal menambahkan ke keranjang",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/keranjang/checkout-preview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan ringkasan untuk checkout dari keranjang",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keranjang"
                ],
                "summary": "Preview checkout",
                "responses": {
                    "200": {
                        "description": "Ringkasan checkout",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Keranjang kosong",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Gagal membuat preview checkout",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/keranjang/clear": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus semua item dari keranjang user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keranjang"
                ],
                "summary": "Kosongkan keranjang",
                "responses": {
                    "200": {
                        "description": "Keranjang berhasil dikosongkan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Gagal mengosongkan keranjang",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/keranjang/count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan jumlah total item di keranjang user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keranjang"
                ],
                "summary": "Hitung jumlah item keranjang",
                "responses": {
                    "200": {
                        "description": "Jumlah item keranjang",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Gagal menghitung jumlah item",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/keranjang/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengupdate jumlah item di keranjang",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keranjang"
                ],
                "summary": "Update item keranjang",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Keranjang",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data update jumlah",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateKeranjangRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item keranjang berhasil diupdate",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Item keranjang tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Stok produk tidak mencukupi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Gagal mengupdate item keranjang",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus item dari keranjang (soft delete dengan mengubah status)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keranjang"
                ],
                "summary": "Hapus item dari keranjang",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Keranjang",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item berhasil dihapus dari keranjang",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ID tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Item keranjang tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Gagal menghapus item",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.CartSummaryResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.KeranjangItemResponse"
                    }
                },
                "total_barang": {
                    "type": "integer",
                    "example": 3
                },
                "total_harga": {
                    "type": "number",
                    "example": 150000
                },
                "total_item": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "controllers.CreateKeranjangRequest": {
            "type": "object",
            "required": [
                "jumlah",
                "produk_id"
            ],
            "properties": {
                "jumlah": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 2
                },
                "produk_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "controllers.KeranjangItemResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "harga_satuan": {
                    "type": "number",
                    "example": 50000
                },
                "jumlah": {
                    "type": "integer",
                    "example": 2
                },
                "keranjang_id": {
                    "type": "integer",
                    "example": 1
                },
                "produk": {
                    "$ref": "#/definitions/models.Produk"
                },
                "produk_id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "subtotal": {
                    "type": "number",
                    "example": 100000
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "controllers.UpdateKeranjangRequest": {
            "type": "object",
            "required": [
                "jumlah"
            ],
            "properties": {
                "jumlah": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 3
                }
            }
        },
...
    }
}